Jest testien alustus backendiin
