+++ server/package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "NODE_ENV=development node index.js",
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest --verbose",
     "start:prod": "NODE_ENV=production node index.js",
     "start:test": "NODE_ENV=test node index.js",
     "start:mongo": "cd ../mongo && docker-compose up && cd ../mongo && docker-compose down"
@@ -19,5 +19,9 @@
     "express": "^4.17.1",
     "mongoose": "^5.11.15",
     "mongoose-unique-validator": "^2.0.3"
+  },
+  "devDependencies": {
+    "jest": "^26.6.3",
+    "supertest": "^6.1.3"
   }
 }
