+++ server/requests/all_products.rest
@@ -1,0 +1,1 @@
+get  http://localhost:3001/api/products
