+++ server/tests/kierratysavustin_api.test.js
@@ -1,0 +1,16 @@
+const mongoose = require('mongoose')
+const supertest = require('supertest')
+const app = require('../app')
+
+const api = supertest(app)
+
+test('products are returned as json', async () => {
+  await api
+    .get('/api/products')
+    .expect(200)
+    .expect('Content-Type', /application\/json/)
+})
+
+afterAll(() => {
+  mongoose.connection.close()
+})
